pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "planning-mode-eval"
version = "0.1.0"
description = "Evaluation pipeline for Claude Code Plan Mode"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.18.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
plan-eval = "src.run_pipeline:main"

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-dir]
"" = "."
